+++ src/Valuify.Tests/Snippets/Declarations/Nested.InClass.Expected.cs
@@ -33,8 +33,7 @@ internal static partial class Nested
                                     return false;
                                 }
 
-                                return global::System.Collections.Generic.EqualityComparer<T1>.Default.Equals(left.Age, right.Age)
-                                    && global::System.Collections.Generic.EqualityComparer<T2>.Default.Equals(left.Name, right.Name);
+                                return left.Equals(right);
                             }
                         }
                     }
@@ -62,7 +61,7 @@ internal static partial class Nested
                     {
                         partial class Inner<T2>
                         {
-                            public sealed override bool Equals(object other)
+                            public override bool Equals(object other)
                             {
                                 return Equals(other as Inner<T2>);
                             }
@@ -121,7 +120,18 @@ internal static partial class Nested
                         {
                             public bool Equals(Inner<T2> other)
                             {
-                                return this == other;
+                                if (ReferenceEquals(this, other))
+                                {
+                                    return true;
+                                }
+                
+                                if (ReferenceEquals(other, null))
+                                {
+                                    return false;
+                                }
+                
+                                return global::System.Collections.Generic.EqualityComparer<T1>.Default.Equals(Age, other.Age)
+                                    && global::System.Collections.Generic.EqualityComparer<T2>.Default.Equals(Name, other.Name);
                             }
                         }
                     }
@@ -149,7 +159,7 @@ internal static partial class Nested
                     {
                         partial class Inner<T2>
                         {
-                            public sealed override int GetHashCode()
+                            public override int GetHashCode()
                             {
                                 return global::Valuify.Internal.HashCode.Combine(Age, Name);
                             }
@@ -209,7 +219,7 @@ internal static partial class Nested
                     {
                         partial class Inner<T2>
                         {
-                            public sealed override string ToString()
+                            public override string ToString()
                             {
                                 return string.Format("Inner { Age = {0}, Name = {1} }", Age, Name);
                             }