+++ src/Valuify/Strategies/EquatableStrategy.cs
@@ -6,6 +6,8 @@ using Valuify.Model;
 internal sealed class EquatableStrategy
     : IStrategy
 {
+    private const string Conditional = "\r\n            && ";
+
     /// <inheritdoc/>
     public IEnumerable<Source> Generate(Subject subject)
     {
@@ -34,16 +36,43 @@ internal sealed class EquatableStrategy
 
     private static Source GenerateImplementation(Subject subject)
     {
+        string conditions = "true";
+
+        if (subject.Properties.Count > 0)
+        {
+            IEnumerable<string> properties = subject.Properties
+                .Select(property => $"{GetComparer(property)}.Default.Equals({property.Name}, other.{property.Name})");
+
+            conditions = string.Join(Conditional, properties);
+        }
+
         string code = $$"""
             partial class {{subject.Qualification}}
             {
                 public bool Equals({{subject.Qualification}} other)
                 {
-                    return this == other;
+                    if (ReferenceEquals(this, other))
+                    {
+                        return true;
+                    }
+            
+                    if (ReferenceEquals(other, null))
+                    {
+                        return false;
+                    }
+            
+                    return {{conditions}};
                 }
             }
             """;
 
         return new Source(code, $"{nameof(IEquatable<Subject>)}.{nameof(Equals)}");
+    }
+
+    private static string GetComparer(Property property)
+    {
+        return property.IsSequence
+            ? $"global::Valuify.Internal.SequenceEqualityComparer"
+            : $"global::System.Collections.Generic.EqualityComparer<{property.Type}>";
     }
 }